Use Clang formatting for native C++ code
The Clang-Format:
https://clang.llvm.org/docs/ClangFormatStyleOptions.html
is used in this project to ensure the same common standard is kept for
the C++ development.

The standard is used mostly by Microsoft but a bit modified.
It is a mix of Microsoft and GNU mostly with custom settings.

The .clang-format configuration file is placed in the root of Windows
solution in windows/ directory to ensure that it will be used for
Windows native side only and won't reach the JS side.