Rename Fingerprint to SignIn
The react-native-windows-hello library will be composed of two main
modules: windows passport and windows signing-in
Those two will be implemented by native modules of this library.

The fingerprint native module should be responsible for all signing-in
options, not just the fingerprint. For that reason it is necessary to
rename it to SignIn, so it will contain the implementation of all
Windows Hello signing-in possibilities.