Use callback mechanism for async SignIn methods
When using the asynchronous methods, the result is unknown at the time
of resolving the promise, for this reason it is impossible to deduce
whether promise should be resolved or rejected.
To make it more user-friendly the async methods are returning the
callback, which is only to notifiy the caller that the method has been
launched successfully.

Only after having the notification the result can be fetched by using
the `getLastResult` methods for both availability and scan results.