Native API improvements - new approach to `SignIn` results (#12)
* Rename Fingerprint to SignIn

The react-native-windows-hello library will be composed of two main
modules: windows passport and windows signing-in
Those two will be implemented by native modules of this library.

The fingerprint native module should be responsible for all signing-in
options, not just the fingerprint. For that reason it is necessary to
rename it to SignIn, so it will contain the implementation of all
Windows Hello signing-in possibilities.

* Add Async methods to the native API

The async methods of SignIn native module use the coroutines to pull the
value of windows hello sign-in procedure.
They both returns the promise, as every native call is asynchronous in
that way, but the `Async` methods are launched in yet separate threads.

* Change 'fingerprint' occurences to 'biometric'

* Use callback separately with promise handled methods

* Reject the promise on an error

If any error occurs during the verification or device availability check
the Promise will be rejected.
This is by checking the status equal to 0 of the status enumaration.

* Correct the success message for verification

* Use callback mechanism for async SignIn methods

When using the asynchronous methods, the result is unknown at the time
of resolving the promise, for this reason it is impossible to deduce
whether promise should be resolved or rejected.
To make it more user-friendly the async methods are returning the
callback, which is only to notifiy the caller that the method has been
launched successfully.

Only after having the notification the result can be fetched by using
the `getLastResult` methods for both availability and scan results.