Refactor the promise-based methods to return integer
Previously returning message converted with WinRTHelper class, the
consent verification and verifier status methods are now returning the
enumerated native value of the `Windows::Security::Credentials::UI`
which corresponds to the status value.
These values however return these values casted to integers, so that it
is possible to pass them directly to `React::JSValue` constructor.
This value should be then converted back into enum object of SignIn
module.

NOTE:
No matter of the result of biometric action, the `Promise` will be
resolved as if it was successeded - means that even if the enum value is
different then `Available` or `Verified` the promise will still not
reject it.
This change is due to no comparison condition when getting the native
value, so that the promise rejection can only be made in case of
internal error or any other failure then related to app's user action.