Adjust `SignInModule` methods for new native API
Both `requestConsentVerification` and `getDeviceStatus` still return the
promise, but they are now adjusted to the fact that this Promise is
always resolved if there's no internal error noticed.

`requestConsentVerification` is the reimplemented `getConsentMessage`
which (same as `getDeviceStatus`) does not return the message now, but
the result object of one of enumerations.