Adjust the messages for statuses according to docs
It turns out that there's a mismatch between the docs examples and
documentation for WinRT text strings for the verification results.
This commit adjusts those to the messages specified in the WinRT
documentation.
The only exception left is the text for the
userConsentVerificationResult.Verified
which was left at the state of code docs.
The reason for that is that the docs are saying that the Fingerprint
verification was successful while this library allows to use other
biometric devices (e.g. camera).