New JS API - results returned as enum objects (#23)
* Install the latest of Jest and Babel for syntax

* Enumerate statuses for consent verification and availability

* Update and reconfigure the preset-env for Jest

* Move the enumerated statuses and conversion function to Statuses

* Cover Statuses with unit tests for enum values correctness

* Refactor the promise-based methods to return integer

Previously returning message converted with WinRTHelper class, the
consent verification and verifier status methods are now returning the
enumerated native value of the `Windows::Security::Credentials::UI`
which corresponds to the status value.
These values however return these values casted to integers, so that it
is possible to pass them directly to `React::JSValue` constructor.
This value should be then converted back into enum object of SignIn
module.

NOTE:
No matter of the result of biometric action, the `Promise` will be
resolved as if it was successeded - means that even if the enum value is
different then `Available` or `Verified` the promise will still not
reject it.
This change is due to no comparison condition when getting the native
value, so that the promise rejection can only be made in case of
internal error or any other failure then related to app's user action.

* Adjust `SignInModule` methods for new native API

Both `requestConsentVerification` and `getDeviceStatus` still return the
promise, but they are now adjusted to the fact that this Promise is
always resolved if there's no internal error noticed.

`requestConsentVerification` is the reimplemented `getConsentMessage`
which (same as `getDeviceStatus`) does not return the message now, but
the result object of one of enumerations.

* Export results enums as objects

* Adjust the messages for statuses according to docs

It turns out that there's a mismatch between the docs examples and
documentation for WinRT text strings for the verification results.
This commit adjusts those to the messages specified in the WinRT
documentation.
The only exception left is the text for the
userConsentVerificationResult.Verified
which was left at the state of code docs.
The reason for that is that the docs are saying that the Fingerprint
verification was successful while this library allows to use other
biometric devices (e.g. camera).