Update package.json with latest react native
Both react-native and react-native-windows are updated to the version
0.67.1 and 0.67.2 respectively.
This change is applied only to the package.json so only the frameworks
are updated here.

NOTE: this commit also replaces yarn with npm as a deps tool.
npm is still used most commonly and can be used by everyone, while not
everyone has Yarn installed.